Refatoração do service, controller e tratamento de ClassCastException para eventos programados e não programados
