+++ event_service/README.md
@@ -1,0 +1,29 @@
+<h1 align="center">Urban Mobility Management System</h1>
+
+----
+<p align="center">
+<a href="https://github.com/FellipeToledo/Urban-Mobility-Management-System">
+  <img src="https://raw.githubusercontent.com/FellipeToledo/files/refs/heads/main/u-m-m-high-resolution-logo-transparent.png" alt="umms-logo" width="200px" height="60px"/></a>
+  <br>
+  <em>O Urban Mobility Management System é uma plataforma que gerencia eventos de trânsito em tempo real
+    <br> otimizando o tráfego e melhorando a mobilidade urbana.</em>
+  <br>
+</p>
+
+<p align="center">
+  <a href="#Key-features">Key Features</a> •
+  <a href="#how-to-use">How To Use</a> •
+  <a href="#credits">Credits</a> •
+  <a href="#related">Related</a> •
+  <a href="#license">License</a>
+</p>
+
+---
+
+<div align="center">
+
+
+
+![screenshot](https://raw.githubusercontent.com/FellipeToledo/files/refs/heads/main/github-desktop.svg) [@FellipeToledo](https://github.com/FellipeToledo) &nbsp;&middot;&nbsp;   ![screenshot](https://raw.githubusercontent.com/FellipeToledo/files/refs/heads/main/linkedin-outlined.svg) [@FellipeToledo](https://www.linkedin.com/in/fellipetoledo/) &nbsp;&middot;&nbsp;
+
+</div>
