+++ event_service/src/main/java/com/azvtech/event_service/controller/EventController.java
@@ -1,0 +1,4 @@
+package com.azvtech.event_service.controller;
+
+public class EventController {
+}
