+++ event_service/src/main/java/com/azvtech/event_service/enums/Severity.java
@@ -1,0 +1,4 @@
+package com.azvtech.event_service.enums;
+
+public enum Severity {
+}
