+++ event_service/src/main/java/com/azvtech/event_service/model/Event.java
@@ -1,0 +1,4 @@
+package com.azvtech.event_service.model;
+
+public class Event {
+}
