+++ event_service/src/main/java/com/azvtech/event_service/model/UnscheduledEvent.java
@@ -1,0 +1,4 @@
+package com.azvtech.event_service.model;
+
+public class UnscheduledEvent {
+}
