+++ event_service/src/main/java/com/azvtech/event_service/repository/EventRepository.java
@@ -1,0 +1,4 @@
+package com.azvtech.event_service.repository;
+
+public interface EventRepository {
+}
