+++ event_service/src/main/java/com/azvtech/event_service/service/EventService.java
@@ -1,0 +1,4 @@
+package com.azvtech.event_service.service;
+
+public class EventService {
+}
