--- src/main/java/com/azvtech/mobility/urbanmobilitymanagement/UrbanMobilityManagementSystemApplication.java
@@ -1,13 +1,0 @@
-package com.azvtech.mobility.urbanmobilitymanagement;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UrbanMobilityManagementSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UrbanMobilityManagementSystemApplication.class, args);
-	}
-
-}
