--- src/main/java/com/azvtech/mobility/urbanmobilitymanagement/controller/EventController.java
@@ -1,98 +1,0 @@
-package com.azvtech.mobility.urbanmobilitymanagement.controller;
-
-import com.azvtech.mobility.urbanmobilitymanagement.model.Event;
-import com.azvtech.mobility.urbanmobilitymanagement.service.EventService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/v1/events")
-public class EventController {
-
-    private final EventService eventService;
-
-    public EventController(EventService eventService) {
-        this.eventService = eventService;
-    }
-
-    @Operation(summary = "Buscar todos os eventos")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Lista de eventos recuperada com sucesso"),
-            @ApiResponse(responseCode = "404", description = "Nenhum evento encontrado")
-    })
-    @GetMapping("/events")
-    public ResponseEntity<List<Event>> getAllEvents() {
-        List<Event> events = eventService.findAllEvents();
-        return events.isEmpty()
-                ? ResponseEntity.noContent().build()
-                : ResponseEntity.ok(events);
-    }
-
-    @Operation(summary = "Buscar um evento pelo ID")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Evento encontrado com sucesso"),
-            @ApiResponse(responseCode = "404", description = "Evento não encontrado")
-    })
-    @GetMapping("/{id}")
-    public ResponseEntity<Event> getEventById(
-            @Parameter(description = "ID do evento a ser buscado")
-            @PathVariable Long id) {
-        return eventService.findEventById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
-    }
-
-    @Operation(summary = "Criar um novo evento")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "Evento criado com sucesso"),
-            @ApiResponse(responseCode = "400", description = "Dados inválidos para criar o evento")
-    })
-    @PostMapping
-    public ResponseEntity<Event> createEvent(
-            @Valid
-            @RequestBody Event event) {
-        Event newEvent = eventService.createEvent(event);
-        return ResponseEntity
-                .status(HttpStatus.CREATED)
-                .header("Location", "/api/resource/" + newEvent.getId())
-                .body(newEvent);
-    }
-
-    @Operation(summary = "Atualizar um evento existente")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Evento atualizado com sucesso"),
-            @ApiResponse(responseCode = "404", description = "Evento não encontrado")
-    })
-    @PutMapping("/{id}")
-    public ResponseEntity<Event> updateEvent(
-            @Parameter(description = "ID do evento a ser atualizado")
-            @PathVariable Long id,
-            @Valid
-            @RequestBody Event updatedEvent) {
-        return eventService.updateEvent(id, updatedEvent)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
-    }
-
-    @Operation(summary = "Deletar um evento")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Evento deletado com sucesso"),
-            @ApiResponse(responseCode = "404", description = "Evento não encontrado")
-    })
-    @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteEvent(
-            @Parameter(description = "ID do evento a ser deletado")
-            @PathVariable Long id) {
-        return eventService.deleteEvent(id) ?
-                ResponseEntity.noContent().build() :
-                ResponseEntity.notFound().build();
-    }
-}
