--- src/main/java/com/azvtech/mobility/urbanmobilitymanagement/repository/EventRepository.java
@@ -1,7 +1,0 @@
-package com.azvtech.mobility.urbanmobilitymanagement.repository;
-
-import com.azvtech.mobility.urbanmobilitymanagement.model.Event;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EventRepository extends JpaRepository<Event, Long> {
-}
