--- src/main/resources/application.yaml
@@ -1,21 +1,0 @@
-#Swagger
-springdoc:
-  api-docs:
-    path: /v3/api-docs
-  swagger-ui:
-    path: /swagger-ui.html
-  info:
-    title: Urban Mobility Management API
-    description: API para gerenciamento de eventos e mobilidade urbana.
-    version: v1.0
-
-
-#Actuator
-management:
-  endpoints:
-    web:
-      exposure:
-        include: health, info, metrics
-  endpoint:
-    health:
-      show-details: always
