--- src/test/java/com/azvtech/mobility/urbanmobilitymanagement/UrbanMobilityManagementSystemApplicationTests.java
@@ -1,13 +1,0 @@
-package com.azvtech.mobility.urbanmobilitymanagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UrbanMobilityManagementSystemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
